refactor(bridge): extract SearchMessageHandler
